feat: 모듈별 코드 개선 및 LangSmith 트레이싱 연동
- workflow.py: 문서 로드 및 검색 로직 개선
- nodes.py: QueryIntentParser, RFPAnswerGenerator 개선
- state.py: 데이터 클래스 수정
- parsers: HWP 로더 개선
- retrievers: VectorStore 임베딩 로직 수정
- prompts: 템플릿 업데이트
- utils/config: LangSmith 환경 변수 추가

+++ src/graph/state.py
@@ -16,7 +16,7 @@ class OrgInfo:
     """기관 정보를 저장하는 데이터 클래스."""
     name: str
     amount: str = ""
-    amount_numeric: int = 0
+    amount_numeric: float = 0
     project_name: str = ""
     summary: str = ""
     open_date: str = ""
